Improve config handling
Many projects need a specific config object (usually the global one)
populated for project code that relies on it. Additionally, we want
to make the code used to initialize the object overrideable so projects
can use their existing code for that. The primary benefit is that
project code has sane defaults for config file locations so in many
cases it won't be necessary to explicitly specify the location. A
very simple default implementation is also provided so the module
still works out of the box too.

Change-Id: If8ce7202bd73793688a0446f069bb4e03547c99e