tighten API for main()
Have the caller set up any of their options before passing a
ConfigOpts instance to main() and then let main() take over processing
from there.

Change the check callback argument to take an UpgradeCommands instance
so in the future if there are more commands we can get the callbacks
out of it ourself.

Add a default_config_files option to main() so that applications can
pass locations oslo.config won't find by default.

Change 'category' to 'command'.

Use a closure in _register_cli_options instead of a separate function
that requires a partial.

Change-Id: Ic4f25fd23d424acfcfe185a85bb63abf7d25c39e
Signed-off-by: Doug Hellmann <doug@doughellmann.com>