Move displayed checker title to be in variable
Title of check command displayed to user is now moved
to be defined as atribute of UpgradeCommands class.
This will allow to override this command for projects which
inherits from this class.

This will be usefull for Neutron to use same class in new
sanity-check tool which will be implemented in same way as
upgrade-check is done.

Change-Id: I8218b21a219db3a3198750fd98033aff244b7002
Related-Bug: #1809878